docs: improve Node.js snippets (#586)
Improves the following aspects of Node.js snippets:
* Adds strict mode to all scripts
* Adds missing error handling to ILP samples
* Removes redundant socket.on('data') listener
* Fixes early socket.destroy() call in ILP samples:
  destroy call didn't wait for all records to be
  flushed into the kernel buffer
* Writes to socket now wait for the previous write
  to complete, hence writes have backpressure

+++ snippets/insert/ilp.js
@@ -1,3 +1,5 @@
+"use strict"
+
 const net = require("net")
 
 const client = new net.Socket()
@@ -12,18 +14,30 @@ function run() {
       `test,location=uk temperature=11.4 ${Date.now() * 1e6}`,
     ]
 
-    rows.forEach((row) => {
-      client.write(`${row}\n`)
-    })
+    function write(idx) {
+      if (idx === rows.length) {
+        client.destroy()
+        return
+      }
 
-    client.destroy()
+      client.write(rows[idx] + "\n", (err) => {
+        if (err) {
+          console.error(err)
+          process.exit(1)
+        }
+        write(++idx)
+      })
+    }
+
+    write(0)
   })
 
-  client.on("data", function (data) {
-    console.log("Received: " + data)
+  client.on("error", (err) => {
+    console.error(err)
+    process.exit(1)
   })
 
-  client.on("close", function () {
+  client.on("close", () => {
     console.log("Connection closed")
   })
 }
